rubik/functions/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
MarkupSafe==1.0
pip==18.0
requests==2.21.0
setuptools==40.2.0
Werkzeug==0.14.1
wheel==0.31.1
google-cloud-pubsub